atualizado a estrutura, usado o Fluent API na context e mapeada a DTO
