+++ Desenvolvimento/BackEnd/TechVagasAPI/TechVagasAPI/Context/AppDbContext.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.EntityFrameworkCore;
-using TechVagasAPI.Models;
+using TechVagasAPI.Models.Entities;
 
 namespace TechVagasAPI.Context
 {
@@ -9,5 +9,13 @@ namespace TechVagasAPI.Context
         { }
 
         public DbSet<CargoModel>? Cargos { get; set; }
-    }
+
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            modelBuilder.Entity<CargoModel>().HasKey(c => c.CargoId);
+            modelBuilder.Entity<CargoModel>().Property(c => c.Descricao).HasMaxLength(100).IsRequired();
+            modelBuilder.Entity<CargoModel>().Property(c => c.Tipo).HasMaxLength(50).IsRequired();
+        }
+
+	}
 }
