--- Desenvolvimento/BackEnd/TechVagasAPI/TechVagasAPI/Dtos/CargoDto.cs
@@ -1,6 +1,0 @@
-﻿namespace TechVagasAPI.Dtos
-{
-	public class CargoDto
-	{
-	}
-}
