--- Desenvolvimento/BackEnd/TechVagasAPI/TechVagasAPI/Models/AdministradorModel.cs
@@ -1,14 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace TechVagasAPI.Models
-{
-	public class AdministradorModel : PessoaModel
-	{
-		[Key]
-		public int AdministradorId { get; set; }
-		public DateTime DataCadastro { get; set; }
-		public string? Status { get; set; }
-
-
-	}
-}
