--- Desenvolvimento/BackEnd/TechVagasAPI/TechVagasAPI/Models/CordenadorEstagioModel.cs
@@ -1,12 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace TechVagasAPI.Models
-{
-	public class CordenadorEstagioModel : PessoaModel
-	{
-		[Key]
-		public int CordenadorEstagioId { get; set; }
-		public DateTime DataCadastro { get; set; }
-		public string? Status { get; set; }
-	}
-}
