+++ Desenvolvimento/BackEnd/TechVagasAPI/TechVagasAPI/Models/Entities/AdministradorModel.cs
@@ -1,0 +1,14 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace TechVagasAPI.Models.Entities
+{
+    public class AdministradorModel : PessoaModel
+    {
+        [Key]
+        public int AdministradorId { get; set; }
+        public DateTime DataCadastro { get; set; }
+        public string? Status { get; set; }
+
+
+    }
+}
