+++ Desenvolvimento/BackEnd/TechVagasAPI/TechVagasAPI/Models/Entities/CargoModel.cs
@@ -1,10 +1,7 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace TechVagasAPI.Models
+﻿namespace TechVagasAPI.Models.Entities
 {
     public class CargoModel
     {
-        [Key]
         public int CargoId { get; set; }
         public string? Descricao { get; set; }
         public string? Tipo { get; set; }
