+++ Desenvolvimento/BackEnd/TechVagasAPI/TechVagasAPI/Models/Entities/ConcedenteModel.cs
@@ -1,0 +1,12 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace TechVagasAPI.Models.Entities
+{
+    public class ConcedenteModel : PessoaModel
+    {
+        [Key]
+        public int ConcedenteId { get; set; }
+        public string? ResponsavelEstagio { get; set; }
+
+    }
+}
